# Improvements to performance of accessing aggregate fields (closes #17)