Its not perfect, but a 25,000x improvement is acceptable.
For a simple test of accessing nested fields performance went from:
```
BenchmarkTools.Trial: 
  memory estimate:  2.95 KiB
  allocs estimate:  121
  --------------
  minimum time:     62.624 μs (0.00% GC)
  median time:      64.228 μs (0.00% GC)
  mean time:        67.337 μs (2.16% GC)
  maximum time:     14.649 ms (99.11% GC)
  --------------
  samples:          10000
  evals/sample:     1
```
to this:
```
BenchmarkTools.Trial: 
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     2.774 ns (0.00% GC)
  median time:      2.779 ns (0.00% GC)
  mean time:        2.784 ns (0.00% GC)
  maximum time:     12.078 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1000
```